Add [magnum].client_microversion
The [magnum].client_microversion allows to set a specific microversion
to use when interacting with Magnum's API endpoints. The test case
TestCluster.test_create_cluster_with_zero_nodes will only pass when the
microversion is set to 1.10[0] which it's not Ussuri's default.

Ussuri is EOL upstream, although having the chance to run tempest
against a downstream supported version lowers the barrier for distro
fixes.

[0] https://opendev.org/openstack/magnum/commit/0e6d17893

Change-Id: I59d027ddc82999f41cfbffcdceb055b7c29dad42